+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 * Enabled Magento `2.4` support in `composer.json`
+* Added category image size view model
 
 ## [1.0.1] 2020-05-29
 * Removed invalid enter in changelog file